1.0.0-rewrite: bridge release for DL4J 1.0.0-SNAPSHOT
This is a bridge release branch before complete Java package renamespacing.
Maven groupId is fully migrated to org.eclipse.deeplearning4j; Java source
packages (org.deeplearning4j, org.nd4j, org.datavec) remain unchanged for now.

Changes across all modules:
- Java 11 minimum (was Java 8), using maven-compiler-plugin release=11
- Snapshot repo moved to https://central.sonatype.com/repository/maven-snapshots/
- CUDA backend updated to nd4j-cuda-12.9-platform (was 10.2)
- All module versions normalized to 1.0.0-SNAPSHOT
- API fixes: int[] -> long[] for ArgMax, ListBuilder import path
- New examples: SameDiff LLM/VLM/audio, OmniHub, GGML, attention, capsule net,
  advanced training (LoRA, distillation, mixed precision), and more
- Updated README.md and CLAUDE.md documenting bridge release status

+++ samediff-examples/src/main/java/org/nd4j/examples/samediff/customizingdl4j/layers/L2NormalizeLambdaLayer.java
@@ -33,7 +33,7 @@ import org.nd4j.autodiff.samediff.SameDiff;
  */
 public class L2NormalizeLambdaLayer extends SameDiffLambdaLayer {
 
-    private int[] dimensions;
+    private long[] dimensions;
 
     /**
      *
@@ -42,7 +42,7 @@ public class L2NormalizeLambdaLayer extends SameDiffLambdaLayer {
      *                   For RNNs, this would also be dimension 1 (to normalize each time step separately)
      *                   For CNNs, this would be dimensions 1, 2 and 3
      */
-    public L2NormalizeLambdaLayer(int... dimensions){
+    public L2NormalizeLambdaLayer(long... dimensions){
         this.dimensions = dimensions;
     }
 
@@ -60,11 +60,11 @@ public class L2NormalizeLambdaLayer extends SameDiffLambdaLayer {
     }
 
     //Getters and setters for JSON serialization
-    public int[] getDimensions(){
+    public long[] getDimensions(){
         return dimensions;
     }
 
-    public void setDimensions(int[] dimensions){
+    public void setDimensions(long[] dimensions){
         this.dimensions = dimensions;
     }
 }
